decisions/assets/ADR-003/examples/python/requirements.txt
PyJWT==2.8.0
requests==2.31.0
